Introduced oqdict
